feat: add solution for "Next Greater Node In Linked List" problem
Introduce a new Java class to implement the solution for the
"Next Greater Node In Linked List" problem, which includes the
definition of the ListNode class and a method stub for
nextLargerNodes. This sets the foundation for solving the problem
of finding the next greater node for each node in a linked list.