feat: add _1002_FindCommonCharacters class to implement common character finder
This commit introduces a new class that will be used to find common
characters in an array of strings. The initial structure includes
a main method for testing and a placeholder for the solution method.
This sets the foundation for solving the problem as described in
the provided comments and examples.