feat: add NextPermutation class and initial structure for solving the problem
This commit introduces the `_31_NextPermutation` class in two packages,
`org.qinlinj.algocamp.day2` and `org.qinlinj.leetcode.editor.en`, to
provide a solution for the "Next Permutation" problem. The class includes
a placeholder for the `nextPermutation` method and a main method for
testing purposes. This sets the foundation for implementing the
algorithm to find the next lexicographical permutation of an array.